fix(parler): cast bfloat16 audio to float32 before numpy conversion
numpy can't convert bfloat16 tensors directly (TypeError: Got unsupported
ScalarType BFloat16). Call .float() before .numpy() on the generated audio
tensor in both baseline and TurboQuant sweep scripts.

+++ speech-tts-improvements/parler/scripts/01_generate_baseline.py
@@ -81,7 +81,7 @@ for voice_name, voice_desc in VOICES.items():
             torch.cuda.synchronize()
         gen_time = time.time() - t_gen
 
-        audio_arr = audio.cpu().numpy().squeeze().astype(np.float32)
+        audio_arr = audio.float().cpu().numpy().squeeze().astype(np.float32)
         audio_duration = len(audio_arr) / model.config.sampling_rate
 
         out_path = os.path.join(OUT_DIR, f"{key}.wav")
